Fix: `fastn build` failure in CI
Renaming the urls to match the filename somehow makes `fastn build`
work. The last CI run was done about an year ago so it's very hard to
find what broke this build.

Trying to create a min. reproducer out of this package also does not
produce this bug correctly.

The best compromise here is to just change the url. Since a min.
reproducer could not be created, a fastn bug is not filed at the time of
writing this commit.